pyproject.toml
[tool.poetry]
name = "gancio_telegram_bot"
version = "0.1.0"
description = "A Telegram bot to interact with Gancio instances"
authors = ["Gianluca Morcaldi <bendico765@gmail.com>"]
license = "AGPL-3.0-only"
readme = "README.md"
homepage = "https://github.com/bendico765/gancio_telegram_bot"
repository = "https://github.com/bendico765/gancio_telegram_bot"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "20.0a4"
aiosqlite = "^0.17.0"
aiohttp-client-cache = "^0.7.3"
gancio-requests = "^0.1.2"
python-dotenv = "^0.21.0"
pandas = "^1.4.4"
matplotlib = "^3.5.3"
seaborn = "^0.12.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
